Move models to separate files; fix circular imports
