+++ src/edclasses/__init__.py
@@ -1,1 +1,1 @@
-from .classes import *
+from .classes import System, OrbitalStation, Faction, FactionBranch
