+++ src/edclasses/tests/test_relations.py
@@ -1,4 +1,4 @@
-from edclasses import OneToManyRelation
+from ..utils import OneToManyRelation
 
 
 class Parent:
